Avoid login issue in deployment E2E tests (#2065)
## Motivation for the change, related issues

The deployment E2E tests are currently broken. WP admin email
verification for an old WP 6.5 build appears to be causing a login
failure error during the deployment tests.

An example of a failing screenshot comparison due to failed login is
[here](https://github.com/WordPress/wordpress-playground/actions/runs/12232536614/job/34118248072#step:7:446).

## Implementation details

This PR disables login to avoid problems with admin email verification on an old, pre-installed WP build. With this change, we are able to re-enable more deployment e2e tests.

## Testing Instructions (or ideally a Blueprint)

- CI