[CORS Proxy] Support chunked encoding when running in Apache/Nginx/etc (#2114)
Adds support for `transfer-encoding: chunked` on Apache, Nginx, and
other web servers.

#2077 added support for `transfer-encoding: chunked` in a way that works
on a local PHP dev server. The proxy manually chunks the output bytes
and outputs each chunk's header, separator, body, and trailer.

However, it doesn't work on playground.wordpress.net because the web
server there handles the chunked encoding on its own. The bytes echoed
by cors-proxy.php are treated as body bytes, which messes up the
response body.

This PR restricts the manual chunking to a local CLI dev server.

## Testing instructions

Run cors-proxy.php in Apache or so and confirm that requesting resources
served with chunked encoding works with this patch but not without it.
One such URL is https://adamadam.blog/feed/.