Fix scp error in web app deployment (#2155)
## Motivation for the change, related issues

We're currently experiencing an scp error when attempting to transfer
`mime-types.json` as a single file to a the destination
`'~/website-deployment/'`.

https://github.com/WordPress/wordpress-playground/actions/runs/13481525026/job/37667340559#step:7:16325

It seems like the home dir resolution is failing altogether or
inconsistent between scp and regular ssh login. This was working before.
There may have been a change in the hosting environment that
inadvertently affected this script.

The previous step copies a directory to the same destination, and that
is working fine.

## Implementation details

As a workaround, this PR just copies `mime-types.json` into the local
directory that is successfully copied to the server. This way, the file
is transferred to the same location, and we avoid this single-file
weirdness with scp.

## Testing Instructions (or ideally a Blueprint)

- CI
- Try workflow after merging. The workflow is already currently broken,
and it is faster to test this way.